Seperate batch.go generic batch test coverage from BatchPPD test coverage into batch_test
