+++ batchPPD_test.go
@@ -102,221 +102,6 @@ func TestBatchODFIIDMismatch(t *testing.T) {
 	}
 }
 
-func TestBatchNumberMismatch(t *testing.T) {
-	mockBatch := mockBatchPPD()
-	mockBatch.GetControl().BatchNumber = 2
-	if err := mockBatch.Validate(); err != nil {
-		if e, ok := err.(*BatchError); ok {
-			if e.FieldName != "BatchNumber" {
-				t.Errorf("%T: %s", err, err)
-			}
-		} else {
-			t.Errorf("%T: %s", err, err)
-		}
-	}
-}
-
-func TestCreditBatchisBatchAmount(t *testing.T) {
-	mockBatch := NewBatchPPD()
-	mockBatch.SetHeader(mockBatchHeader())
-	e1 := mockEntryDetail()
-	e1.TransactionCode = 22
-	e1.Amount = 100
-	e2 := mockEntryDetail()
-	e2.TransactionCode = 22
-	e2.Amount = 100
-	mockBatch.AddEntry(e1)
-	mockBatch.AddEntry(e2)
-	if err := mockBatch.Create(); err != nil {
-		t.Errorf("%T: %s", err, err)
-	}
-
-	mockBatch.GetControl().TotalCreditEntryDollarAmount = 1
-	if err := mockBatch.Validate(); err != nil {
-		if e, ok := err.(*BatchError); ok {
-			if e.FieldName != "TotalCreditEntryDollarAmount" {
-				t.Errorf("%T: %s", err, err)
-			}
-		} else {
-			t.Errorf("%T: %s", err, err)
-		}
-	}
-}
-
-func TestSavingsBatchisBatchAmount(t *testing.T) {
-	mockBatch := NewBatchPPD()
-	mockBatch.SetHeader(mockBatchHeader())
-	e1 := mockEntryDetail()
-	e1.TransactionCode = 32
-	e1.Amount = 100
-	e2 := mockEntryDetail()
-	e2.TransactionCode = 37
-	e2.Amount = 100
-	mockBatch.AddEntry(e1)
-	mockBatch.AddEntry(e2)
-	if err := mockBatch.Create(); err != nil {
-		t.Errorf("%T: %s", err, err)
-	}
-
-	mockBatch.GetControl().TotalDebitEntryDollarAmount = 1
-	if err := mockBatch.Validate(); err != nil {
-		if e, ok := err.(*BatchError); ok {
-			if e.FieldName != "TotalDebitEntryDollarAmount" {
-				t.Errorf("%T: %s", err, err)
-			}
-		} else {
-			t.Errorf("%T: %s", err, err)
-		}
-	}
-}
-
-func TestBatchisEntryHash(t *testing.T) {
-	mockBatch := mockBatchPPD()
-	mockBatch.GetControl().EntryHash = 1
-	if err := mockBatch.Validate(); err != nil {
-		if e, ok := err.(*BatchError); ok {
-			if e.FieldName != "EntryHash" {
-				t.Errorf("%T: %s", err, err)
-			}
-		} else {
-			t.Errorf("%T: %s", err, err)
-		}
-	}
-}
-
-func TestBatchDNEMismatch(t *testing.T) {
-	mockBatch := NewBatchPPD()
-	mockBatch.SetHeader(mockBatchHeader())
-	ed := mockEntryDetail()
-	ed.AddAddenda(mockAddenda())
-	ed.AddAddenda(mockAddenda())
-	mockBatch.AddEntry(ed)
-	mockBatch.Create()
-
-	mockBatch.GetHeader().OriginatorStatusCode = 1
-	mockBatch.GetEntries()[0].TransactionCode = 23
-	if err := mockBatch.Validate(); err != nil {
-		if e, ok := err.(*BatchError); ok {
-			if e.FieldName != "OriginatorStatusCode" {
-				t.Errorf("%T: %s", err, err)
-			}
-		} else {
-			t.Errorf("%T: %s", err, err)
-		}
-	}
-}
-
-func TestBatchTraceNumberNotODFI(t *testing.T) {
-	mockBatch := mockBatchPPD()
-	mockBatch.GetEntries()[0].setTraceNumber(12345678, 1)
-	if err := mockBatch.Validate(); err != nil {
-		if e, ok := err.(*BatchError); ok {
-			if e.FieldName != "ODFIIdentificationField" {
-				t.Errorf("%T: %s", err, err)
-			}
-		} else {
-			t.Errorf("%T: %s", err, err)
-		}
-	}
-}
-
-func TestBatchEntryCountEquality(t *testing.T) {
-	mockBatch := NewBatchPPD()
-	mockBatch.SetHeader(mockBatchHeader())
-	e := mockEntryDetail()
-	a := mockAddenda()
-	e.AddAddenda(a)
-	mockBatch.AddEntry(e)
-	if err := mockBatch.Create(); err != nil {
-		t.Errorf("%T: %s", err, err)
-	}
-
-	mockBatch.GetControl().EntryAddendaCount = 1
-	if err := mockBatch.Validate(); err != nil {
-		if e, ok := err.(*BatchError); ok {
-			if e.FieldName != "EntryAddendaCount" {
-				t.Errorf("%T: %s", err, err)
-			}
-		} else {
-			t.Errorf("%T: %s", err, err)
-		}
-	}
-}
-
-func TestBatchAddendaIndicator(t *testing.T) {
-	mockBatch := mockBatchPPD()
-	mockBatch.GetEntries()[0].AddAddenda(mockAddenda())
-	mockBatch.GetEntries()[0].AddendaRecordIndicator = 0
-	mockBatch.GetControl().EntryAddendaCount = 2
-	if err := mockBatch.Validate(); err != nil {
-		if e, ok := err.(*BatchError); ok {
-			if e.FieldName != "AddendaRecordIndicator" {
-				t.Errorf("%T: %s", err, err)
-			}
-		} else {
-			t.Errorf("%T: %s", err, err)
-		}
-	}
-}
-
-func TestBatchIsAddendaSeqAscending(t *testing.T) {
-	mockBatch := NewBatchPPD()
-	mockBatch.SetHeader(mockBatchHeader())
-	ed := mockEntryDetail()
-	ed.AddAddenda(mockAddenda())
-	ed.AddAddenda(mockAddenda())
-	mockBatch.AddEntry(ed)
-	mockBatch.Create()
-
-	mockBatch.GetEntries()[0].Addendum[0].(*Addenda).SequenceNumber = 2
-	mockBatch.GetEntries()[0].Addendum[1].(*Addenda).SequenceNumber = 1
-	if err := mockBatch.Validate(); err != nil {
-		if e, ok := err.(*BatchError); ok {
-			if e.FieldName != "SequenceNumber" {
-				t.Errorf("%T: %s", err, err)
-			}
-		} else {
-			t.Errorf("%T: %s", err, err)
-		}
-	}
-}
-
-func TestBatchIsSequenceAscending(t *testing.T) {
-	mockBatch := mockBatchPPD()
-	e3 := mockEntryDetail()
-	e3.TraceNumber = 1
-	mockBatch.AddEntry(e3)
-	mockBatch.GetControl().EntryAddendaCount = 2
-	if err := mockBatch.Validate(); err != nil {
-		if e, ok := err.(*BatchError); ok {
-			if e.FieldName != "TraceNumber" {
-				t.Errorf("%T: %s", err, err)
-			}
-		} else {
-			t.Errorf("%T: %s", err, err)
-		}
-	}
-}
-
-func TestBatchAddendaTraceNumber(t *testing.T) {
-	mockBatch := mockBatchPPD()
-	mockBatch.GetEntries()[0].AddAddenda(mockAddenda())
-	if err := mockBatch.Create(); err != nil {
-		t.Errorf("%T: %s", err, err)
-	}
-
-	mockBatch.GetEntries()[0].Addendum[0].(*Addenda).EntryDetailSequenceNumber = 99
-	if err := mockBatch.Validate(); err != nil {
-		if e, ok := err.(*BatchError); ok {
-			if e.FieldName != "TraceNumber" {
-				t.Errorf("%T: %s", err, err)
-			}
-		} else {
-			t.Errorf("%T: %s", err, err)
-		}
-	}
-}
-
 func TestBatchBuild(t *testing.T) {
 	mockBatch := NewBatchPPD()
 	header := NewBatchHeader()
